Fix lit tests to work with multiple sources

+++ src/backend/main.cpp
@@ -143,7 +143,7 @@ static std::vector<IncludeEntity> get_all_includes_from_source(
         if (std::regex_search(lines[i_line], match, include_regex)) 
         {
             bool match_is_at_start_line = true;
-            for (int i = 0; i < lines[i_line].size(); i++) {
+            for (size_t i = 0; i < lines[i_line].size(); i++) {
                 if (lines[i_line][i] != match.str()[i]) {
                     match_is_at_start_line = false;
                     break;
